% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translationCode.R
\name{freesurferToMatrix}
\alias{freesurferToMatrix}
\title{Read freesurfer files into a Matrix}
\usage{
freesurferToMatrix(freesurferFileNames, valuetype = "thickness")
}
\arguments{
\item{valuetype}{for example "thickness"}
}
\value{
A matrix containing the freesurfer data.
}
\description{
Read freesurfer files into rows of a matrix, given a valuetype i.e. something
stored in a freesurfer vertex or surface file.
}
\examples{

# TODO

}
\author{
Avants BB
}
\seealso{
\code{\link{imagesToMatrix}, \link{getMask}}
}

