% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translationCode.R
\name{freesurferToMatrix}
\alias{freesurferToMatrix}
\title{Read freesurfer files into a Matrix}
\usage{
freesurferToMatrix(freesurferFileNames)
}
\arguments{
\item{freesurferFileNames}{a list of filenames pointing to freesurfer data}
}
\value{
A matrix containing the freesurfer data.
}
\description{
Read freesurfer files into rows of a matrix, given a valuetype i.e. something
stored in a freesurfer vertex or surface data file.
}
\examples{

if ( freesurfer::have_fs( ) ) {
  fname = file.path( freesurfer::fs_subj_dir(), "bert", "surf", "lh.thickness")
  thfns = c( fname, fname )
  mat = freesurferToMatrix( thfns )
  }
\dontrun{
xs = scale( x , center = TRUE, scale = FALSE )
smat = sparseDistanceMatrix( xs, k = 10, kmetric = "covariance",
  eps = 1e-06, mypkg = "rflann")
mydecom = irlba::irlba( smat, nv = 5 )
myproj = x \%*\% mydecom$v
pheatmap::pheatmap( cor( myproj ) )
hist( abs( mydecom$v[,1] )[ abs( mydecom$v[,1] )>1.e-4] )
# same kind of thing with sparseDistanceMatrixXY
}

}
\author{
Avants BB
}

